refactor(server): extract ResetState into its own module
``main.rs`` was 2020 lines. Roughly 300 of those were the ``ResetState``
struct (22 state handles plus three runtime handles), its
``reset_service``/``reset`` methods and the test that exercises them.
That's a cohesive unit with a single responsibility (letting the admin
reset endpoint clear per-service state) and it has no reason to sit in
the entry-point file.

This commit moves ``ResetState`` and its impls to a new
``fakecloud-server::reset`` module, makes the two methods ``pub(crate)``,
imports ``ResetState`` back into ``main.rs``, and keeps the
``rds_instance_response`` test in ``main.rs`` (since it doesn't belong
to reset).

``main.rs`` goes from 2020 to 1712 lines. No behavior change; 1102
unit tests still pass.
